requirements.txt
apache-libcloud=3.7.0
diagrams=0.23.3
